cleaned up code, created component and style directories
