+++ web/App.js
@@ -1,13 +1,16 @@
 import "react-native-gesture-handler";
 import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
-import { StyleSheet } from "react-native";
 import MainNavigator from "./src/components/MainNavigator";
+import { store } from "./src/constants/store";
+import { Provider } from 'react-redux';
 
 export default function App() {
   return (
-    <NavigationContainer>
-      <MainNavigator />
+    <NavigationContainer >
+      <Provider store={store}>
+          <MainNavigator />
+      </Provider>
     </NavigationContainer>
   );
 }
