+++ web/src/components/Front.jsx
@@ -2,33 +2,29 @@ import { createStackNavigator } from "@react-navigation/stack";
 import React from "react";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { ROUTES } from "../constants";
-import ItemCards from "./ItemCards";
-import ItemDescript from "./ItemDescript";
+import Slides from "../pages/Slides";
+import ItemDescription from "../pages/ItemDescription";
 import Navbar from "./Navbar";
 
-const Front = () => {
-  const Stack = createStackNavigator();
+const Stack = createStackNavigator();
 
+export default Front = () => {
   return (
-    <>
       <SafeAreaView>
         <Navbar />
-      </SafeAreaView>
-      <Stack.Navigator>
-        <Stack.Screen
-          name={ROUTES.ITEM_CARDS}
-          component={ItemCards}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name={ROUTES.ITEM_DESCRIPTION}
-          component={ItemDescript}
-          options={{ headerShown: false }}
-        />
-      
-      </Stack.Navigator>
-    </>
+        <Stack.Navigator>
+          <Stack.Screen
+            name={ROUTES.SLIDES}
+            component={Slides}
+            options={{ headerShown: false }}
+          />
+          <Stack.Screen
+            name={ROUTES.ITEM_DESCRIPTION}
+            component={ItemDescription}
+            options={{ headerShown: false }}
+          />
+        </Stack.Navigator>
+    </SafeAreaView>
   );
 };
 
-export default Front;
