--- web/src/components/HeaderBar.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Text } from "react-native";
-
-const HeaderBar = () => {
-  return <Text>write header here?</Text>;
-};
-
-export default HeaderBar;
