--- web/src/components/Signup.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Text } from "react-native";
-
-const Signup = () => {
-  return <Text>Signup</Text>;
-};
-
-export default Signup;
