+++ web/src/constants/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit';
+import searchReducer from './searchSlice';
+
+
+export const store = configureStore({
+    reducer:{
+        search: searchReducer
+    }
+})
