+++ web/src/styles/ItemCard.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  imageBackground: {
+    flex: 1,
+  },
+  itemText: {
+    color: "black",
+    fontSize: 18,
+    fontWeight: "bold",
+  },
+  renderItemContainer: {
+    backgroundColor: "#FFF",
+    flex: 1,
+  },
+});
